[KBE-50] Moved auth into package and deleted sout statements
